v3.2(/v3.1.12):Multi criteria for impact detection/rating, GUI/messages/log improvments

+++ DeTeCt-MFC/dtcas3.cpp
@@ -117,7 +117,7 @@ void read_autostakkert_config_line(std::string line, std::string *filename, std:
 		while (line.find(',') != std::string::npos) replace(line, ",", ".");
 		std::istringstream ss(line);
 		ss >> x >> y;
-		cm_list->push_back(cv::Point(round(x), round(y)));
+		cm_list->push_back(cv::Point((int)round(x), (int)round(y)));
 	}
 }
